+++ src/main/scala/ai/privado/languageEngine/java/passes/read/DatabaseReadUtility.scala
@@ -1,0 +1,152 @@
+/*
+ * This file is part of Privado OSS.
+ *
+ * Privado is an open source static code analysis tool to discover data flows in the code.
+ * Copyright (C) 2022 Privado, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * For more information, contact support@privado.ai
+ *
+ */
+
+package ai.privado.languageEngine.java.passes.read
+
+import ai.privado.cache.{RuleCache, TaggerCache}
+import ai.privado.dataflow.Dataflow
+import ai.privado.model.InternalTag
+import ai.privado.utility.SQLParser
+import ai.privado.utility.Utilities._
+import io.joern.dataflowengineoss.queryengine.EngineContext
+import io.shiftleft.codepropertygraph.generated.nodes.{CfgNode, Expression, TypeDecl}
+import overflowdb.BatchedUpdate.DiffGraphBuilder
+import io.shiftleft.semanticcpg.language._
+import io.joern.dataflowengineoss.queryengine.{EngineConfig, EngineContext}
+import io.joern.dataflowengineoss.language._
+import io.shiftleft.codepropertygraph.generated.Cpg
+
+object DatabaseReadUtility {
+
+  val selectRegexPattern = "(?i)(\")?\\s{0,5}select\\s+.*"
+  val fromRegexPattern   = "(?i)(\")?\\s{0,5}from\\s+.*"
+
+  def processDBReadNode(
+    builder: DiffGraphBuilder,
+    taggerCache: TaggerCache,
+    classTableMapping: Map[String, TypeDecl],
+    cpg: Cpg,
+    node: Expression
+  ) = {
+    val sensitiveClassesWithMatchedRules = taggerCache.typeDeclMemberCache
+    val sensitiveClasses                 = taggerCache.typeDeclMemberCache.keys.l
+    val query                            = extractSQLForConcatenatedString(node.code)
+    val result                           = SQLParser.parseSQL(query)
+
+    result match {
+      case Some(value) =>
+        // Match classes which end with tableNameRegex
+        val tableName      = value._1.toLowerCase
+        val tableNameRegex = s"(?i).*$tableName".r
+        val columns        = value._2
+
+        val sensitiveMemberRuleIds = {
+          if (classTableMapping.contains(tableName) && sensitiveClasses.contains(classTableMapping(tableName).fullName))
+            sensitiveClassesWithMatchedRules(classTableMapping(tableName).fullName).keys.l
+          else
+            sensitiveClasses.find(s => s.matches(tableNameRegex.regex)) match {
+              case Some(value) => sensitiveClassesWithMatchedRules(value).keys.l
+              case None        => List.empty
+            }
+        }
+
+        if (columns.length == 1 && columns(0) == "*") {
+          if (sensitiveMemberRuleIds.nonEmpty)
+            sensitiveMemberRuleIds.foreach(ruleId => addTagsToNode(ruleId, node, builder))
+          else {
+            /* Run dataflow and verify the data-elements read from the call,
+                Ex - resultSet = statement.executeQuery("SELECT * FROM mytable");
+                // Loop through the result set and print out each row
+                while (resultSet.next()) {
+                    int id = resultSet.getInt("id");
+                    String firstName = resultSet.getString("name");
+                    int age = resultSet.getInt("age");
+                    System.out.println("ID: " + id + ", Name: " + firstName + ", Age: " + age)
+                }
+             */
+            val dataElementSinks =
+              Dataflow
+                .getSources(cpg)
+                .filterNot(_.isMember)
+                .map(_.asInstanceOf[CfgNode])
+                .l
+            implicit val engineContext: EngineContext =
+              EngineContext(config = EngineConfig(4))
+            val readFlow = dataElementSinks.reachableByFlows(node).l
+            if (readFlow.nonEmpty) {
+              // As a flow is present from Select query to a Data element we can say, the data element is read from the query
+              readFlow
+                .flatMap(_.elements.last.tag.value("Data.Sensitive.*"))
+                .value
+                .foreach(ruleId => addTagsToNode(ruleId, node, builder))
+            }
+          }
+        } else {
+          if (sensitiveMemberRuleIds.nonEmpty)
+            sensitiveMemberRuleIds
+              .filter(ruleId => isColumnNameMatchingWithRule(ruleId, columns))
+              .foreach(ruleId => addTagsToNode(ruleId, node, builder))
+          else
+            RuleCache.getRule.sources
+              .filter(rule => isColumnNameMatchingWithRule(rule.id, columns))
+              .foreach(rule => addTagsToNode(rule.id, node, builder))
+        }
+      case None => ()
+    }
+
+  }
+
+  /** Return True if any column name matches the pattern
+    *
+    * @param ruleId
+    * @param columns
+    * @return
+    */
+  def isColumnNameMatchingWithRule(ruleId: String, columns: Array[String]): Boolean = {
+    val pattern = RuleCache.getRuleInfo(ruleId).get.combinedRulePattern.r
+    columns.map(pattern.matches).foldLeft(false)(_ || _)
+  }
+
+  def addTagsToNode(ruleId: String, node: Expression, builder: DiffGraphBuilder) = {
+    storeForTag(builder, node)(InternalTag.VARIABLE_REGEX_LITERAL.toString)
+    addRuleTags(builder, node, RuleCache.getRuleInfo(ruleId).get)
+  }
+
+  def extractSQLForConcatenatedString(sqlQuery: String): String = {
+    var query = sqlQuery
+      .stripPrefix("\"")
+      .stripSuffix("\"")
+      .split("\\\"\\s*\\+\\s*\\\"") // Splitting the query on '+' operator and joining back to form complete query
+      .map(_.stripMargin)
+      .mkString("")
+
+    // Add `select *` to queries which are `from users`
+    if (query.matches(fromRegexPattern))
+      query = "select * " + query
+
+    val pattern =
+      "(?i)SELECT\\s(.*?)\\sFROM\\s(.*?)(`.*?`|\".*?\"|'.*?'|\\w+)".r // Pattern to fetch the SELECT statement from the query
+    pattern.findFirstIn(query).getOrElse("")
+  }
+
+}
